# trip-builder/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Trip Builder API",
        "description": "Simple trip builder api",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1/api/v1",
            "description": "api server"
        }
    ],
    "paths": {
...
        "/flights": {
            "get": {
                "tags": [
                    "Flights"
                ],
                "summary": "Get list of flights",
                "description": "Returns list of flights",
                "operationId": "getFlightsList",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Flights"
                ],
                "summary": "Store flight",
                "description": "Create new flight",
                "operationId": "postFlight",
                "requestBody": {
                    "description": "Flight object",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Flight"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/flights/{flight}": {
            "get": {
                "tags": [
                    "Flights"
                ],
                "summary": "Get flight information",
                "description": "Returns flight data",
                "operationId": "getFlightById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Flight id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "put": {
                "tags": [
                    "Flights"
                ],
                "summary": "Update flight",
                "description": "update an flight",
                "operationId": "putFlight",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Flight id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Flight object",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Flight"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Flights"
                ],
                "summary": "Delete flight",
                "description": "delete an flight",
                "operationId": "deleteFlight",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Flight id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/regions": {
            "get": {
                "tags": [
                    "Regions"
                ],
                "summary": "Get list of regions",
                "description": "Returns list of regions",
                "operationId": "getRegionsList",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Regions"
                ],
                "summary": "Store region",
                "description": "Create new region",
                "operationId": "postRegion",
                "requestBody": {
                    "description": "Region object",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Region"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/regions/{region}": {
            "get": {
                "tags": [
                    "Regions"
                ],
                "summary": "Get region information",
                "description": "Returns region data",
                "operationId": "getRegionById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Region id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "put": {
                "tags": [
                    "Regions"
                ],
                "summary": "Update region",
                "description": "update an region",
                "operationId": "putRegion",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Region id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Region object",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Region"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Regions"
                ],
                "summary": "Delete region",
                "description": "delete an region",
                "operationId": "deleteRegion",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Region id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Flight": {
                "type": "object",
                "allOf": [
                    {
                        "required": [
                            "code"
                        ],
                        "properties": {
                            "number": {},
                            "airline_id": {},
                            "departure_airport_id": {},
                            "departure_time": {},
                            "arrival_airport_id": {},
                            "arrival_time": {},
                            "price": {}
                        },
                        "type": "object"
                    }
                ]
            },
            "Region": {
                "type": "object",
                "allOf": [
                    {
                        "required": [
                            "code"
                        ],
                        "properties": {
                            "code": {},
                            "name": {},
                            "country_id": {}
                        },
                        "type": "object"
                    }
                ]
            },
...
        },
        "securitySchemes": {
            "Authorization": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}